Create iOS target
This is the initial check-in of code to allow SS to run on iOS. This is a pair of new targets, kpx_cpu_ios and SheepShaveriOS for app itself. This works to a first approximation. A build of SDL2, as SDL2.xcframework, is also checked in and allows building and targeting to any simulator or device, Intel or Apple Silicon.

SS looks in the Documents directory (the one shared with the Finder) for the ROM file (named Mac OS ROM, not included) and a boot volume (named MacOS9.dsk, also not included). This is in no way ready for prime time, but it can boot and run. Pairing a Bluetooth keyboard to the device is needed, as there isn't a software keyboard (yet). Using a mouse with an iPad is supported.