Be nicer about choosing the ROM file
We now have a directed strategy for finding and loading ROM files. Essentially, if the "rom" preference is valid, we use that. Otherwise, if there is but one recognizable ROM file, we use that. If there are more than one ROM file, we bring up the prefs and ask the user to choose. And if there are none, we inform the user and give them the opportunity to retry, thereby giving them the chance to copy in a ROM file without having to quit and relaunch.