Add Frame Rate control to AV prefs
Added a stepper-based frame rate control to the AV pane with a 60 Hz default. Note that dynamic ( == "as fast as possible") is not available, because this makes SS use 100% of a CPU core. That's more or less OK on a computer, but iOS casts a jaundiced eye on processes that take up that much CPU on a given thread and after 15 minutes it will kill SS. We also tweaked the automatic ROM picker a bit.