Networking works now.
Slirp can't read from /etc/resolv.conf because it's outside the permitted parts of the file system, so we came up with another way for it to get the DNS, and now Ethernet networking works. Also make sure the disk list properly updates after the user creates a new disk.