Got build going on Xcode13
Modules in Obj-C and C++ are automatically enabled on iOS projects and this results in an overly aggressive include use of system headers, even when double quotes are used in the include line. This should force use of the local version of the file, but Xcode 13 will pick up the system version.